+++ src/components/Page/Page.module.css.d.ts
@@ -2,6 +2,8 @@ declare const styles: {
   readonly "bar": string;
   readonly "barTitle": string;
   readonly "body": string;
+  readonly "bodySkeleton": string;
+  readonly "bodyState": string;
   readonly "compact": string;
   readonly "header": string;
   readonly "large": string;